+++ features/agreement/step_definitions/agreement-creation.ts
@@ -102,25 +102,6 @@ Given(
 );
 
 Given(
-  "{string} ha una richiesta di fruizione in stato {string} per quell'e-service",
-  async function (consumer: TenantType, agreementState: string) {
-    assertContextSchema(this, {
-      eserviceId: z.string(),
-      descriptorId: z.string(),
-      token: z.string(),
-    });
-    const token = getToken(this.tokens, consumer, "admin");
-    this.agreementId =
-      await dataPreparationService.createAgreementWithGivenState(
-        token,
-        agreementState,
-        this.eserviceId,
-        this.descriptorId
-      );
-  }
-);
-
-Given(
   "{string} non possiede uno specifico attributo certificato",
   async function (tenantType: TenantType) {
     assertContextSchema(this, {
